+++ AssistantKit/Classes/Bundle.swift
@@ -8,57 +8,71 @@
 
 import Foundation
 
-/**
- Used for getting current build information.
- 
- **Getting bundle version**
- 
- Return bundle version string:
- 
- ```
- Bundle.bundleVersion // "3.1.0" i.e.
- ```
- 
- Return bundle short version string:
- 
- ```
- Bundle.bundleShortVersion // "3.1" i.e
- ```
- 
- Getting bundle identifier is easy too:
- 
- ```
- Bundle.bundleIdentifier // com.apple.example i.e.
- ```
- 
- **Default directories**
- 
- As it's often required to get access to Document or Cache directories inside your bundle
- I've added few methods that makes it easy to work with it.
- 
- Working with .DocumentDirectory:
- 
- ```
- Bundle.documentsDirectoryURL                       // Return NSURL to .DocumentDirectory
- Bundle.documentsDirectoryPath                      // Return path string to .DocumentDirectory
- Bundle.filePathInDocumentsDirectory("test.data")   // Return path to file inside .DocumentDirectory
- ```
- 
- Working with .CachesDirectory:
-
- ```
- Bundle.cachesDirectoryURL                          // Return NSURL to .CachesDirectory
- Bundle.cachesDirectoryPath                         // Return path string to .CachesDirectory
- Bundle.filePathInCachesDirectory("user.cache")     // Return path to file inside .CachesDirectory
- ```
-
- */
+/// Used for getting current build information.
+///
+/// **Getting bundle version**
+///
+/// Return bundle version string:
+///
+/// ```
+/// Bundle.bundleVersion // "3.1.0" i.e.
+/// ```
+///
+/// Return bundle short version string:
+///
+/// ```
+/// Bundle.bundleShortVersion // "3.1" i.e
+/// ```
+///
+/// Getting bundle identifier is easy too:
+///
+/// ```
+/// Bundle.bundleIdentifier // com.apple.example i.e.
+/// ```
+///
+/// **Default directories**
+///
+/// It's often required to get access to Document or Cache directories inside your bundle
+/// I've added few methods that makes it easy to work with it.
+///
+/// Working with .DocumentDirectory:
+///
+/// ```
+/// Bundle.documentsDirectoryURL                       // Return NSURL to .DocumentDirectory
+/// Bundle.documentsDirectoryPath                      // Return path string to .DocumentDirectory
+/// Bundle.filePathInDocumentsDirectory("test.data")   // Return path to file inside .DocumentDirectory
+/// ```
+///
+/// Working with .CachesDirectory:
+///
+/// ```
+/// Bundle.cachesDirectoryURL                          // Return NSURL to .CachesDirectory
+/// Bundle.cachesDirectoryPath                         // Return path string to .CachesDirectory
+/// Bundle.filePathInCachesDirectory("user.cache")     // Return path to file inside .CachesDirectory
+/// ```
+///
+/// **Working with directories**
+///
+/// There are few helper methods to make access to Documents and Caches directoies easier.
+/// Take a look at code examples:
+///
+/// ```
+/// Bundle.documentsDirectoryURL           // URL to .DocumentDirectory
+/// Bundle.documentsDirectoryPath          // Path to .DocumentDirectory
+/// Bundle.cachesDirectoryURL              // URL to .CachesDirectory
+/// Bundle.cachesDirectoryPath             // Path to .CachesDirectory
+///
+/// let filePath = "directory/filename.txt"
+///
+/// Bundle.filePathInDocumentsDirectory(toFile: filePath)  // Path to file in .DocumentDirectory
+/// Bundle.filePathInCachesDirectory(toFile: filePath)     // Path to file in .CachesDirectory
+///
+/// ```
 public struct Bundle {
 
 }
 
-// MARK: Getting build details information
-
+/// Getting build details information
 extension Bundle {
 
     /// Return bundle version
@@ -77,8 +91,7 @@ extension Bundle {
     }
 }
 
-// MARK: Getting default app folders
-
+/// Getting default app folders
 extension Bundle {
 
     // MARK: .DocumentDirectory