+++ AssistantKit/Classes/Type.swift
@@ -8,13 +8,13 @@
 
 import Foundation
 
-/**
- - Parameter Phone:
- - Parameter Pad:
- - Parameter Pod:
- - Parameter Simulator:
- - Parameter Unknown
- */
+/// Device types
+///
+/// - Parameter phone:
+/// - Parameter pad:
+/// - Parameter pod:
+/// - Parameter simulator:
+/// - Parameter unknown
 public enum Type: String {
     case phone
     case pad
@@ -23,9 +23,7 @@ public enum Type: String {
     case unknown
 }
 
-/**
- Exact device version
- */
+/// Exact device version
 public enum Version: String {
     case phone4
     case phone4S
@@ -39,6 +37,9 @@ public enum Version: String {
     case phoneSE
     case phone7
     case phone7Plus
+    case phone8
+    case phone8Plus
+    case phoneX
 
     case pad1
     case pad2
@@ -63,10 +64,9 @@ public enum Version: String {
 
     case unknown
     
-    /**
-     Return device screen
-     - Seealso: Screen
-     */
+    /// Return device screen
+    ///
+    /// - seealso: Screen
     public var screen: Screen {
         switch self {
         case .podTouch1,
@@ -76,6 +76,7 @@ public enum Version: String {
              .phone4,
              .phone4S:
             return .inches_3_5
+            
         case .podTouch5,
              .podTouch6,
              .phone5,
@@ -83,19 +84,28 @@ public enum Version: String {
              .phone5S,
              .phoneSE:
             return .inches_4_0
+            
         case .phone6,
              .phone6S,
-             .phone7:
+             .phone7,
+             .phone8:
             return .inches_4_7
+            
         case .phone6Plus,
              .phone6SPlus,
-             .phone7Plus:
+             .phone7Plus,
+             .phone8Plus:
             return .inches_5_5
+
+        case .phoneX:
+            return.inches_5_8
+            
         case .padMini,
              .padMini2,
              .padMini3,
              .padMini4:
             return .inches_7_9
+            
         case .pad1,
              .pad2,
              .pad3,
@@ -103,26 +113,25 @@ public enum Version: String {
              .padAir,
              .padAir2:
             return .inches_9_7
+            
         case .padPro:
             return .inches_12_9
-        case .simulator,
-             .unknown:
+            
+        default:
             return .unknown
         }
     }
 }
 
-/**
- Used to determinate device type
- */
+/// Used to determinate device type
 extension Device {
 
-    static fileprivate func versionCode() -> String {
+    /// Return raw device version code string or empty string if any problem appears.
+    static public var versionCode: String {
         var systemInfo = utsname()
         uname(&systemInfo)
 
-        if let
-            info = NSString(bytes: &systemInfo.machine, length: Int(_SYS_NAMELEN), encoding: String.Encoding.ascii.rawValue),
+        if  let info = NSString(bytes: &systemInfo.machine, length: Int(_SYS_NAMELEN), encoding: String.Encoding.ascii.rawValue),
             let code = String(validatingUTF8: info.utf8String!)
         {
             return code
@@ -131,12 +140,11 @@ extension Device {
         return ""
     }
 
-    /**
-     Return device type
-     - Seealso: Type
-    */
+    /// Return device type
+    ///
+    /// - seealso: Type
     static public var type: Type {
-        let versionCode = Device.versionCode()
+        let versionCode = Device.versionCode
 
         switch versionCode {
         case "iPhone3,1", "iPhone3,2", "iPhone3,3",
@@ -146,7 +154,8 @@ extension Device {
              "iPhone6,1", "iPhone6,2",
              "iPhone7,2", "iPhone7,1",
              "iPhone8,1", "iPhone8,2", "iPhone8,4",
-             "iPhone9,1", "iPhone9,2", "iPhone9,3", "iPhone9,4":
+             "iPhone9,1", "iPhone9,2", "iPhone9,3", "iPhone9,4",
+             "iPhone10,1", "iPhone10,2", "iPhone10,3", "iPhone10,4", "iPhone10,5", "iPhone10,6":
             return .phone
 
         case "iPad1,1",
@@ -188,6 +197,11 @@ extension Device {
         return !isPad
     }
 
+    /// Return `true` for iPahoneX
+    static public var isPhoneX: Bool {
+        return isPhone && screen == .inches_5_8
+    }
+
     /// Return `true` for iPadPro
     static public var isPadPro: Bool {
         return isPad && screen == .inches_12_9
@@ -201,7 +215,7 @@ extension Device {
     // MARK: Version
 
     static public var version: Version {
-        let versionCode = Device.versionCode()
+        let versionCode = Device.versionCode
 
         switch versionCode {
         // Phones
@@ -240,6 +254,15 @@ extension Device {
 
         case "iPhone9,2", "iPhone9,4":
             return .phone7Plus
+
+        case "iPhone10,1", "iPhone10,4":
+            return .phone8
+
+        case "iPhone10,2", "iPhone10,5":
+            return .phone8Plus
+
+        case "iPhone10,3", "iPhone10,6":
+            return .phoneX
 
 
         // Pads